% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_groups.R
\name{group_by_groups}
\alias{group_by_groups}
\title{Group Estimates By Groups}
\usage{
group_by_groups(
  object,
  ...,
  col_names = "est",
  group_first = TRUE,
  group_labels = NULL
)
}
\arguments{
\item{object}{A \linkS4class{lavaan} object or the output of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}} or
\code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}.}

\item{...}{Optional arguments to be passed to
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. Ignored if object is an
output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}} or
\code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}.}

\item{col_names}{A vector of the column names in the
parameter estimate tables to be included.
Default is \code{"est"}.}

\item{group_first}{If \code{TRUE}, the columns will be grouped
by groups first and then by
columns in the parameter estimates
tables. Default is \code{TRUE}.}

\item{group_labels}{A character vector of group labels.
Will be assigned to group id = 1, 3, 3,
etc. If not provided. will try to be
retrieved from \code{object} if it is
a \link[lavaan:lavaan-class]{lavaan::lavaan} object.}
}
\value{
A data-frame-like object of the class \code{est_table}.
}
\description{
Groups parameter estimates or other information such as
p-values into a table with groups as columns
and parameters as rows.
}
\examples{

# TODO: Amend
library(lavaan)
set.seed(5478374)
n <- 50
x <- runif(n) - .5
m <- .40 * x + rnorm(n, 0, sqrt(1 - .40))
y <- .30 * m + rnorm(n, 0, sqrt(1 - .30))
dat <- data.frame(x = x, y = y, m = m)
model <-
'
m ~ a*x
y ~ b*m
ab := a*b
'

fit <- sem(model, data = dat, fixed.x = FALSE)
parameterEstimates(fit)
group_by_dvs(fit)
group_by_ivs(fit)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
