% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_models.R
\name{group_by_models}
\alias{group_by_models}
\title{Group Estimates By Models}
\usage{
group_by_models(
  output_list,
  ...,
  col_names = "est",
  group_first = FALSE,
  model_first = TRUE
)
}
\arguments{
\item{output_list}{A named list of \linkS4class{lavaan} objects,
a named list of the output of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}},
or a named list of the output of
\code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}.}

\item{...}{Optional arguments to be passed to
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. Ignored
if the elements in output_list are the results of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}} or
\code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}.}

\item{col_names}{A vector of the column names in the
parameter estimate tables to be included.
Default is \code{"est"}.}

\item{group_first}{If \code{TRUE}, the rows will be grouped
by groups first and then by
parameters.
Ignored if
the model has only one group.
Default is \code{FALSE}.}

\item{model_first}{If \code{TRUE}, the columns will be grouped
by models first and then by
columns in the parameter estimates
tables. Default is \code{TRUE}.}
}
\value{
A data-frame-like object of the class \code{est_table}.
}
\description{
Groups parameter estimates or other information such as
p-values into a table with models as columns.
}
\examples{

# TODO: Amend
library(lavaan)
set.seed(5478374)
n <- 50
x <- runif(n) - .5
m <- .40 * x + rnorm(n, 0, sqrt(1 - .40))
y <- .30 * m + rnorm(n, 0, sqrt(1 - .30))
dat <- data.frame(x = x, y = y, m = m)
model <-
'
m ~ a*x
y ~ b*m
ab := a*b
'

fit <- sem(model, data = dat, fixed.x = FALSE)
parameterEstimates(fit)
group_by_dvs(fit)
group_by_ivs(fit)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
