% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semhelpinghands-package.R
\docType{package}
\name{semhelpinghands-package}
\alias{semhelpinghands}
\alias{semhelpinghands-package}
\title{semhelpinghands: Helper Functions for Structural Equation Modeling}
\description{
An assortment of helper functions for doing structural equation modeling, mainly by 'lavaan' for now. Most of them are time-saving functions for common tasks in doing structural equation modeling and reading the output. This package is not for functions that implement advanced statistical procedures. It is a light-weight package for simple functions to simple tasks conveniently, as few dependencies as possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semhelpinghands/}
  \item Report bugs at \url{https://github.com/sfcheung/semhelpinghands/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

}
\keyword{internal}
