% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizedSolution_boot_ci.R
\name{standardizedSolution_boot_ci}
\alias{standardizedSolution_boot_ci}
\title{Percentile Bootstrap CIs for Standardized Solution}
\usage{
standardizedSolution_boot_ci(object, level = 0.95, type = "std.all", ...)
}
\arguments{
\item{object}{A \linkS4class{lavaan} object, fitted with 'se = "boot"'.}

\item{level}{The level of confidence of the confidence intervals. Default
is .95.}

\item{type}{The type of standard estimates. The same argument of
\code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}, and support all values supported by
\code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}. Default is \code{"std.all"}.}

\item{...}{Other arguments to be passed to \code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}.}
}
\value{
The output of \code{\link[lavaan:standardizedSolution]{lavaan::standardizedSolution()}}, with percentile
bootstrap confidence intervals appended to the right.
}
\description{
It receives a \link[lavaan:lavaan-class]{lavaan::lavaan} object fitted with
bootstrapping standard errors requested and forms the percentile
confidence intervals for the standardized solution.
}
\examples{

library(lavaan)
set.seed(5478374)
n <- 50
x <- runif(n) - .5
m <- .40 * x + rnorm(n, 0, sqrt(1 - .40))
y <- .30 * m + rnorm(n, 0, sqrt(1 - .30))
dat <- data.frame(x = x, y = y, m = m)
model <-
'
m ~ a*x
y ~ b*m
ab := a*b
'

# Should set bootstrap to at least 2000 in real studies
fit <- sem(model, data = dat, fixed.x = FALSE,
           se = "boot",
           bootstrap = 100)
summary(fit)

standardizedSolution_boot_ci(fit)

}
\author{
Shu Fai Cheung (shufai.cheung@gmail.com)
}
