% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_vectors.R
\name{vec_rsquare}
\alias{vec_rsquare}
\alias{vec_sample_vcov}
\alias{vec_sample_var}
\alias{vec_est_var}
\alias{vec_est_se}
\alias{vec_def_se}
\alias{vec_lavTestLRT}
\alias{vec_lavTestScore}
\alias{vec_lavTestWald}
\alias{vec_def_var}
\alias{vec_compRelSEM}
\alias{vector_from_lavaan}
\title{Wrapper Functions to Extract
Information as a Vector}
\usage{
vec_rsquare(object)

vec_sample_vcov(object)

vec_sample_var(object)

vec_est_var(object)

vec_est_se(object)

vec_def_se(object)

vec_lavTestLRT(
  object,
  ...,
  method = "default",
  A.method = "delta",
  scaled.shifted = TRUE,
  H1 = TRUE,
  model.names = NULL
)

vec_lavTestScore(
  object,
  add = NULL,
  release = NULL,
  univariate = TRUE,
  information = "expected"
)

vec_lavTestWald(object, constraints = NULL, prefix = NULL)

vec_def_var(object)

vec_compRelSEM(object, ...)
}
\arguments{
\item{object}{A \code{lavaan}-class
object.}

\item{...}{Additional arguments to
be passed to the original function.}

\item{method}{(Of lavTestLRT.)}

\item{A.method}{(Of lavTestLRT.)}

\item{scaled.shifted}{(Of lavTestLRT.)}

\item{H1}{(Of lavTestLRT.)}

\item{model.names}{(Of lavTestLRT.)}

\item{add}{(Of lavTestScore.)}

\item{release}{(Of lavTestScore.)}

\item{univariate}{(Of lavTestScore.)}

\item{information}{(Of lavTestScore.)}

\item{constraints}{(Of lavTestScore.)}

\item{prefix}{(Of lavTestScore.)}
}
\value{
All of them return a named numeric
vector.
}
\description{
A set of vector functions
to extract information from a
\code{lavaan}-class object and return
a named vector.
}
\details{
This set of wrapper
functions are for functions like
\code{\link[lavaan:bootstrap]{lavaan::bootstrapLavaan()}} that
expects a function that receives
a \code{lavaan}-class object and returns
a vector of values.

Many of the tasks can be performed
by writing our own functions. The
wrapper functions are just time-saver
for common tasks.

The wrapper functions are designed
to be as simple to use as possible,
with as few arguments as possible.
If advanced control is needed, users
are recommended to write their own
wrappers.
}
\section{Functions}{
\itemize{
\item \code{vec_rsquare()}: Get R-squares in a model.

\item \code{vec_sample_vcov()}: Get sample variances and covariances.

\item \code{vec_sample_var()}: Get sample variances.

\item \code{vec_est_var()}: Sampling variances of free parameters.

\item \code{vec_est_se()}: Standard errors of free parameters.

\item \code{vec_def_se()}: Standard errors of user-defined parameters.

\item \code{vec_lavTestLRT()}: Get sample variances.

\item \code{vec_lavTestScore()}: Do score tests.

\item \code{vec_lavTestWald()}: Do a Wald test.

\item \code{vec_def_var()}: Sampling variances of user-defined parameters.

\item \code{vec_compRelSEM()}: Composite reliability.

}}
\examples{
\donttest{
}

}
\seealso{
\code{\link[lavaan:lavInspect]{lavaan::lavInspect()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
